+++ main.ts
@@ -1,6 +1,6 @@
 import { Plugin } from "obsidian";
 
-import { NLTTable } from "src/index";
+import { NLTTable } from "src/NLTTable";
 
 // Remember to rename these classes and interfaces!
